get plots working again, initialize agent location in experiment
